Type compatibility

+++ src/main.rs
@@ -36,7 +36,7 @@ fn main() {
     let s3_addr = S3::create(|ctx| {
         use actix::prelude::*;
         let bucket = env::var("AWS_BUCKET").unwrap();
-        let interval = u16::from_str_radix(&env::var("AWS_INTERVAL").unwrap(), 10).unwrap();
+        let interval = u64::from_str_radix(&env::var("AWS_INTERVAL").unwrap(), 10).unwrap();
 
         ctx.run_interval(Duration::from_secs(interval), |_, ctx| {
             ctx.address().do_send(backends::Flush)
